Preserve AnyEvent watchers/guards
It's a usual pattern in AnyEvent to return a guard that, if it is destroyed,
will deactivate the event.  We now store the results of the todo until we've
gotten notification via our condvar that everything is done.  This ensures
that the event hangs around long enough to be useful.